Neovim設定のリファクタリングと新しいプラグインの追加 - init.lua: require文のスタイルを統一 - base.lua: ambiwidthを'single'に変更 - mapping.lua: キーマッピングの整理とコメントの追加 - lazy_plugins_claudecode.lua: 新しいプラグインの追加 - lazy_plugins_snacks.lua: 新しいプラグインの追加と設定 - lazy_plugins_ui.lua: 条件付きプラグインの設定を追加 - lazy_plugins_novscode.lua: GitHubテーマのコンパイル確認を追加 - lazy_toggle_term.lua: キーマッピングを修正

+++ neovim_config/lua/base.lua
@@ -31,7 +31,7 @@ vim.o.autoread = true
 vim.o.hidden = true
 vim.o.writebackup = false
 vim.o.backup = false
-vim.o.ambiwidth = 'double'
+vim.o.ambiwidth = 'single'
 vim.o.visualbell = true
 vim.o.virtualedit = 'onemore'
